util/grub-mount.c: Improve performance of readdir and getattr
The problem:

fuse_readdir_call_fill() is a callback routine called on every direntry.
fuse_readdir_call_fill() calls grub_file_open() in order to get file size
and other stuff. grub_file_open() does the full path resolution, so it
scans the same directory tree again to locate the file inode.

This makes the O()-complexity of the algorithm to be quadratic of the
number of entries in the directory. Even worse, if an application calls
getattr right next, getattr does same the path resolution one more time.

A separate issue which contributes to the total computational complexity
is symlink resolution. A symlink can possibly trigger any number of
directory rescans, and the total evaluation time is unpredictable.

Solution:

Firstly, don't call grub_file_open() from fuse_readdir_call_fill() at
all. According to the FUSE documentation, the readdir callback routine is
allowed but not obligated to fill the stat information. So, if
grub_dirhook_info data was provided by the grub fs driver, we can pass
the stat info to the kernel. And if the data is not provided, we pass the
file name only.

Secondly, a path lookup cache is implemented on the grub-mount side. It
allows sharing stat info between readdir and getattr invocations and
mostly eliminates need of rescanning the same directories several times
in a row.

The size field added to the struct grub_dirhook_info in order for a fs
driver to be able to provide all needed stat information for
fuse_readdir_call_fill() callback. The code is tested with ext2 fs driver
only. Other drivers should also be modified accordingly, which is not yet
done.

The cache implementation in this commit is quick and dirty and should
probably be rewritten. The lookup is implemented as a simple linear scan
of array. This may impact performance, but at the moment we seem to have
one more performance issue on the grub vfs side, which affects the
performance more compared to the the cache lookup function.